pyproject.toml
[project]
name = "clinical-data-mgmt-snowflake-dbt-demo"
version = "0.1.0"
description = "Clinical Data Management ELT demo on Snowflake with dbt, Prefect, Great Expectations (uv-managed)."
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
  "pandas==2.2.2",
  "python-dotenv==1.0.1",
  "snowflake-connector-python==3.12.0",
  "snowflake-snowpark-python==1.20.0",
  "dbt-core==1.8.6",
  "dbt-snowflake==1.8.3",
  "great-expectations==0.18.12",
  "prefect>=2.19.9,<3.0.0", # ✅ Prefect 2
  "pydantic>=2.7,<3",
  "sqlfluff==3.0.7",
  "black==24.8.0",
  "isort==5.13.2",
  "pytest==8.2.0",
  "pyarrow>=22.0.0",
]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.isort]
profile = "black"

[tool.sqlfluff]
dialect = "snowflake"
exclude_rules = "L031"
